Added some Ttview.js
Added Ttview.js which will display the actual timetable
Timetable.js will be the root component in the Timetable tab of the app.
TtView will be imported into Timetable.js and will be displayed in there.

TtView.js can also be imported into the substitution tab when displaying the timetable for the respective teacher.

Dependencies NOTE:
-We will be using the react-native-elements beta version due to several additional optioins available
-We are using the react-navigation v2 as it has more support at the current stage and v3 has some serious differences compared to v2.