Update TTview.js
Updated the Ttview.js to include map function instead of standard for loop.

Issue for each timing to be in a new row is because of using the cols.push to insert the time text.

Fix this by changing the way to display the array or by seperately displaying times under a different view component.