Update App.js
Added a login page with a close to complete fetch api. Login button currently has no authentication and is meant so for debugging purposes. Image has to be added at the top end of the page but currently adding an Image tag breaks the page to a white blank screen.