The Changes in schema updated
I checked this time, the right schema is up. If it's not lemme know.
Individual teacher's timetable for our class or updated. Their profile is right now not a priority, if you want it populated for testing PM me and I'll do it ASAP. I'll discuss queries for different operations in chat and update you guys.